+++ CHANGELOG.md
@@ -23,6 +23,7 @@ SPDX-License-Identifier: AGPL-3.0-only
 - Increased edge case parameter in `assignzArgBound` in `src/crandall.c` from $10^{16}$ to `DBL_MAX` from `<float.h>` to ensure that the asymptotic method is not used for large $|\nu|$.
 - Added correction term for non-diagonal matrixes in $\nu = d+2 k$ for the regularized Epstein zeta function.
 - Removed special use of asymptotic expansion around $\nu = 2$ and $\nu=4$ which reduces the error from $<10^{-11}$to $<10^{-14}$.
+- Fixed special cases `NU` element of $\{\ldots,-4,-2,0,2,3,4,\ldots\}$ for evaluation of `python examples/python/sem_gaussian_1d.py --nu [NU]`.
 
 ## [0.4.2] - unreleased